backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-slugify==8.0.4
pytest==7.4.4
pytest-asyncio==0.23.7
httpx==0.27.0
python-multipart==0.0.9
